[REFACTOR] Move Favor Room Directory Position for Enhancement Management
